% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jrpg}
\alias{jrpg}
\title{Draw a vector of Polya-Gamma latent random variables with parameters `b` and `z`}
\usage{
jrpg(b, z)
}
\arguments{
\item{b}{vector of parameters for $PG(b,z)$ random variable.}

\item{z}{vector of parameters for $PG(b,z)$ random variable.}
}
\value{
column vector of \code{length(b)} containing $PG(b,z)$ random variates.
}
\description{
Draw a vector of Polya-Gamma latent random variables with parameters `b` and `z`
}
